Added cookie support using Response and Request objects
Setting a cookie in Eupheme can be done using
Response.cookies.set_cookie.

Cookies are handled using the CookieManager class which has support for
normal cookies as well as signed cookies (requires a secret key in the config
file for the application). Cookies are signed using HMAC-SHA-224
(planned to be configurable).

Request objects have a read-only CookieManager instance for accessing
cookies that the client sent along to the server.

If a resource returns something that is not a Response object then that
is passed as data (and later used by faucets and the likes).